Syncing files from `cloud/release/2025.004` to `release/2025.004`

+++ applications/vanilla/models/PostMetaModel.php
@@ -20,7 +20,7 @@ class PostMetaModel extends PipelineModel
      *
      * @param PostFieldModel $postFieldModel
      */
-    public function __construct(private PostFieldModel $postFieldModel)
+    public function __construct(private PostFieldModel $postFieldModel, private PostTypeModel $postTypeModel)
     {
         parent::__construct("postMeta");
     }
@@ -63,15 +63,18 @@ class PostMetaModel extends PipelineModel
      *
      * @param string $recordType
      * @param int $recordID
-     * @param array $postFields
+     * @param array $postMeta
      * @return void
      * @throws \Throwable
      */
-    public function updatePostFields(string $recordType, int $recordID, array $postFields)
+    public function updatePostMeta(string $recordType, int $recordID, array $postMeta)
     {
+        if (empty($postMeta)) {
+            return;
+        }
         try {
             $this->database->beginTransaction();
-            foreach ($postFields as $postFieldID => $postFieldValue) {
+            foreach ($postMeta as $postFieldID => $postFieldValue) {
                 $values = is_array($postFieldValue) ? $postFieldValue : [$postFieldValue];
                 $this->delete([
                     "recordType" => $recordType,
@@ -79,14 +82,11 @@ class PostMetaModel extends PipelineModel
                     "postFieldID" => $postFieldID,
                 ]);
                 foreach ($values as $value) {
-                    $value = $this->normalizeValue($value);
-                    $queryValue = $this->createQueryValue($postFieldID, $value);
                     $this->insert([
                         "recordType" => $recordType,
                         "recordID" => $recordID,
                         "postFieldID" => $postFieldID,
                         "value" => $value,
-                        "queryValue" => $queryValue,
                     ]);
                 }
             }
@@ -95,6 +95,13 @@ class PostMetaModel extends PipelineModel
             $this->database->rollbackTransaction();
             throw $e;
         }
+    }
+
+    public function insert(array $set, array $options = [])
+    {
+        $set["queryValue"] = $this->createQueryValue($set["postFieldID"], $set["value"]);
+        $set["value"] = $this->normalizeValue($set["value"]);
+        return parent::insert($set, $options);
     }
 
     /**
@@ -175,8 +182,18 @@ class PostMetaModel extends PipelineModel
         return $queryValue;
     }
 
-    public function joinPostFields(array &$rowOrRows)
+    /**
+     * Join post fields with an array of discussion records.
+     *
+     * @param array $rowOrRows
+     * @return void
+     * @throws \Exception
+     */
+    public function joinPostMeta(array &$rowOrRows): void
     {
+        if (!PostTypeModel::isPostTypesFeatureEnabled()) {
+            return;
+        }
         reset($rowOrRows);
         $single = is_string(key($rowOrRows));
         if ($single) {
@@ -185,34 +202,28 @@ class PostMetaModel extends PipelineModel
             $rows = &$rowOrRows;
         }
 
-        // First build the query.
-        $recordTypesForLookup = [];
-        $sql = $this->createSql()
-            ->from($this->getTable())
-            ->select()
-            ->select("recordType,'|',recordID", "concat", "recordKey");
-        foreach ($rows as $row) {
-            $sql->orWhere(["recordType" => $row["postTypeID"], "recordID" => $row["discussionID"]]);
-            $recordTypesForLookup[] = $row["postTypeID"];
-        }
+        $discussionIDs = array_column($rows, "discussionID");
+        $postTypeIDs = array_column($rows, "postTypeID");
 
-        // Run the query to get post meta values and index them by recordType (postTypeID) + recordID and postFieldID.
-        $postMetaRecords = $sql->get()->resultArray();
-        $postMetaRecords = ArrayUtils::arrayColumnArrays($postMetaRecords, null, "recordKey");
+        // Run the query to get post meta values and index them by recordID and postFieldID.
+        $postMetaRecords = $this->select(["recordType" => "discussion", "recordID" => $discussionIDs]);
+        $postMetaRecords = ArrayUtils::arrayColumnArrays($postMetaRecords, null, "recordID");
         $postMetaRecords = array_map(fn($record) => $this->condensePostMeta($record), $postMetaRecords);
 
         // Get post field definitions indexed by postTypeID and postFieldID
-        $postFields = $this->postFieldModel->select(["postTypeID" => $recordTypesForLookup, "isActive" => true]);
+        $postFields = $this->postFieldModel->getWhere(["postTypeID" => $postTypeIDs, "isActive" => true]);
         $postFields = ArrayUtils::arrayColumnArrays($postFields, null, "postTypeID");
         $postFields = array_map(fn($field) => array_column($field, null, "postFieldID"), $postFields);
 
+        // Get system fields that serve as catch-all fields for when posts are converted.
+        $systemPostFields = $this->postFieldModel->getWhere(["isActive" => true, "isSystemHidden" => true]);
+        $systemPostFields = array_column($systemPostFields, null, "postFieldID");
+
         // Join the post meta values with the records after normalization and permission checks.
         foreach ($rows as &$row) {
-            $recordKey = $row["postTypeID"] . "|" . $row["discussionID"];
-
-            $recordPostFields = [];
-            foreach ($postMetaRecords[$recordKey] ?? [] as $postFieldID => $value) {
-                $postField = $postFields[$row["postTypeID"]][$postFieldID] ?? null;
+            $recordPostMeta = [];
+            foreach ($postMetaRecords[$row["discussionID"]] ?? [] as $postFieldID => $value) {
+                $postField = $postFields[$row["postTypeID"]][$postFieldID] ?? ($systemPostFields[$postFieldID] ?? null);
                 if (!isset($postField)) {
                     continue;
                 }
@@ -221,17 +232,51 @@ class PostMetaModel extends PipelineModel
                 }
 
                 $value = $this->normalizeValueForOutput($value, $postField["dataType"]);
-                $recordPostFields[$postFieldID] = $value;
+                $recordPostMeta[$postFieldID] = $value;
             }
-            $row["postFields"] = $recordPostFields;
+            $row["postMeta"] = $recordPostMeta;
         }
+    }
+
+    /**
+     * Get post fields for a single discussion record.
+     *
+     * @param int $recordID
+     * @param string $postTypeID
+     * @param array $filters
+     * @return array|null
+     * @throws \Exception
+     */
+    public function getPostMeta(int $recordID, string $postTypeID, array $filters = []): ?array
+    {
+        $postMetaRecords = $this->select(["recordType" => "discussion", "recordID" => $recordID]);
+        $postMetaRecords = $this->condensePostMeta($postMetaRecords);
+
+        $postFields = $this->postFieldModel->getWhere($filters + ["postTypeID" => $postTypeID, "isActive" => true]);
+        $postFields = array_column($postFields, null, "postFieldID");
+
+        // Get system fields that serve as catch-all fields for when posts are converted.
+        $systemPostFields = $this->postFieldModel->getWhere(["isActive" => true, "isSystemHidden" => true]);
+        $systemPostFields = array_column($systemPostFields, null, "postFieldID");
+
+        $recordPostFields = [];
+        foreach ($postMetaRecords as $postFieldID => $value) {
+            $postField = $postFields[$postFieldID] ?? ($systemPostFields[$postFieldID] ?? null);
+            if (!isset($postField)) {
+                continue;
+            }
+
+            $value = $this->normalizeValueForOutput($value, $postField["dataType"]);
+            $recordPostFields[$postFieldID] = $value;
+        }
+        return $recordPostFields;
     }
 
     /**
      * @param $postMeta
      * @return array
      */
-    private function condensePostMeta($postMeta)
+    private function condensePostMeta($postMeta): array
     {
         $postMetaCondensed = [];
         foreach ($postMeta as $metaItem) {
@@ -249,5 +294,105 @@ class PostMetaModel extends PipelineModel
             }
         }
         return $postMetaCondensed;
+    }
+
+    /**
+     * Migrate post fields when moving a record from one post type to another.
+     *
+     * @param int $recordID
+     * @param string $fromPostTypeID
+     * @param string $toPostTypeID
+     * @return void
+     * @throws \Exception
+     */
+    public function movePostMeta(int $recordID, string $fromPostTypeID, string $toPostTypeID): void
+    {
+        // Get the existing post meta for the record.
+        $postMeta = $this->getPostMeta($recordID, $fromPostTypeID);
+
+        // Get the post fields for the destination post type.
+        $postFields = $this->postFieldModel->getWhere(["postTypeID" => $toPostTypeID]);
+        $postFieldsIndexed = array_column($postFields, null, "postFieldID");
+
+        // Get the post fields for the origin post type.
+        $fromPostFields = $this->postFieldModel->getWhere(["postTypeID" => $fromPostTypeID]);
+        $fromPostFieldsIndexed = array_column($fromPostFields, null, "postFieldID");
+
+        // First delete existing post meta.
+        $this->delete(["recordType" => "discussion", "recordID" => $recordID]);
+
+        uksort($postMeta, fn($key) => in_array($key, PostFieldModel::SPECIAL_CATCH_ALL_FIELDS) ? -1 : 0);
+
+        $pendingFields = [];
+        foreach ($postMeta as $postFieldID => $value) {
+            if (isset($postFieldsIndexed[$postFieldID])) {
+                // If the same field is in the destination post type, insert it with the same post field ID.
+                $this->insert([
+                    "recordType" => "discussion",
+                    "recordID" => $recordID,
+                    "postFieldID" => $postFieldID,
+                    "value" => $value,
+                ]);
+                continue;
+            }
+
+            if (in_array($postFieldID, PostFieldModel::SPECIAL_CATCH_ALL_FIELDS)) {
+                // This was already in one of the built-in fields, add the value to the array as-is.
+                $pendingFields[$postFieldID][] = $value;
+                continue;
+            }
+
+            $fromPostFieldVisibility = $fromPostFieldsIndexed[$postFieldID]["visibility"];
+            $fromPostFieldLabel = $fromPostFieldsIndexed[$postFieldID]["label"];
+
+            $postFieldID = self::getPostFieldIDFromVisibility($fromPostFieldVisibility);
+
+            // For all other fields, collect them to process into special built-in post fields.
+            $pendingFields[$postFieldID][] = $fromPostFieldLabel . ": " . self::normalizeValueForDisplay($value);
+        }
+
+        // Insert collected post fields into post meta using special built-in post fields.
+        foreach ($pendingFields as $postFieldID => $fields) {
+            $this->insert([
+                "recordType" => "discussion",
+                "recordID" => $recordID,
+                "postFieldID" => $postFieldID,
+                "value" => implode("\n", $fields),
+            ]);
+        }
+
+        // Delete post field values from origin post type.
+    }
+
+    /**
+     * Convert value for displaying in a text post field.
+     *
+     * @param mixed $value
+     * @return mixed|string
+     */
+    private static function normalizeValueForDisplay(mixed $value): mixed
+    {
+        if (is_array($value)) {
+            return implode(", ", $value);
+        }
+        if (is_bool($value)) {
+            return $value ? "Yes" : "No";
+        }
+        return $value;
+    }
+
+    /**
+     * Get a built-in post field ID for the given visibility.
+     *
+     * @param string $visibility
+     * @return string
+     */
+    private static function getPostFieldIDFromVisibility(string $visibility): string
+    {
+        return match ($visibility) {
+            "public" => PostFieldModel::PUBLIC_DATA_FIELD_ID,
+            "private" => PostFieldModel::PRIVATE_DATA_FIELD_ID,
+            default => PostFieldModel::INTERNAL_DATA_FIELD_ID,
+        };
     }
 }
